+++ ProjectManager/ProjectManager/Controllers/MainViewController.swift
@@ -1,0 +1,39 @@
+//
+//  ProjectManager - ViewController.swift
+//  Created by Jusbug.
+//  Copyright © yagom. All rights reserved.
+// 
+
+import UIKit
+
+class MainViewController: UIViewController {
+    
+    @IBOutlet weak var todoCollectionView: UICollectionView!
+    @IBOutlet weak var doingCollectionView: UICollectionView!
+    @IBOutlet weak var doneCollectionView: UICollectionView!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        configureTitle()
+        
+    }
+    
+    private func configureTitle() {
+        self.navigationItem.title = "Project Manager"
+    }
+    
+    @IBAction func didTapAddButton(_ sender: Any) {
+        guard let secondVC = self.storyboard?.instantiateViewController(withIdentifier: "secondViewController") else { return }
+        
+        secondVC.modalTransitionStyle = .coverVertical
+        secondVC.modalPresentationStyle = .popover
+        
+        if let popover = secondVC.popoverPresentationController {
+            popover.sourceView = self.view
+            popover.sourceRect = CGRect(x: self.view.bounds.midX, y: self.view.bounds.midY, width: 0, height: 0)
+            popover.permittedArrowDirections = []
+        }
+        
+        self.present(secondVC, animated: true, completion: nil)
+    }
+}
